Fix bug
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@11652 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/modes/cutscene_world.cpp
@@ -177,7 +177,7 @@ const std::string& CutsceneWorld::getIdent() const
 /** Update the world and the track.
  *  \param dt Time step size. 
  */
-void CutsceneWorld::update(double dt)
+void CutsceneWorld::update(float dt)
 {
     /*
     {
